refined animation of sessions dragging and dropping
Created a session-tile-draggable component.  This component is a
session-tile that also responds to being dragged by hiding the original
image (leaving only the ghost) and the session name.  This image is
hidden by replacing it with a transparent pixel.

refactored a couple old constructs to have better names (alias of
‘sessions’ for app.model in student, rankedOrNullPrefsArray renamed to
prefsArray)