bug fixes
-Import Text Button now shows a spinner after clicking, and a success
message that confirms the number of students imported and the total
number of students.  (previously, only showed a pop-up, and had no
feedback after clicking button.)

- fixed a regression on the student page. when an admin was updating a
session with a new image, the new image was not updating after a
successful upload.  i fixed that by adding a time stamp to the
imagePath, but that had an unintended side effect of causing images on
the student page to reload whenever they were moved.  I undid the
change to imagePath and created a refreshingImagePath in the
session-pane component to preserve the correct behavior for the admin
while avoiding ugly reloads in the student page.

- fixed a bug where racing to the student metrics panel would cause it
to display incorrect data.