Bug Fix on print pages
last solution to printing a blank page was no good because Ember does
not give us all handlebars features.  So I created a last-item helper
and used it in the rosters so we don’t add a page break to the last
roster.  Working now.  Also added unit testing for the last-item
helper, and started to fix the testing.